Added collider types to differentiate between Triggers, Walls and Platforms.
Triggers: Don't interact with other triggers, but still interact with walls and platforms. Also position gets updated along with the gameobject that holds them.

Platforms: Get ignored by a trigger collider if the trigger collider is in the air, making them invisible when jumping. When falling we stop ignoring them.
This has the effect that we can jump through them when jumping.

Walls: Just static colliders. Nothing interesting about them.